+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Removed
+- Drop deprecated methods from `AbstractDataMapper` (#697)
 - Drop deprecated methods from the testing framework (#693)
 - Drop the legacy non-namespaced class aliases (#692)
 - Drop support for TYPO3 8LTS (#691)